+++ lib/report.js
@@ -4,7 +4,8 @@ const libCoverage = require('istanbul-lib-coverage')
 const libReport = require('istanbul-lib-report')
 const reports = require('istanbul-reports')
 const { readdirSync, readFileSync, statSync } = require('fs')
-const { isAbsolute, resolve, join, relative, extname, dirname } = require('path')
+const { isAbsolute, resolve, extname } = require('path')
+const getSourceMapFromFile = require('./source-map-from-file')
 // TODO: switch back to @c88/v8-coverage once patch is landed.
 const v8toIstanbul = require('v8-to-istanbul')
 const isCjsEsmBridgeCov = require('./is-cjs-esm-bridge')
@@ -130,29 +131,6 @@ class Report {
   }
 
   /**
-   * //TODO: use https://www.npmjs.com/package/convert-source-map
-   * // no need to roll this ourselves this is already in the dep tree
-   * https://sourcemaps.info/spec.html
-   * @param {String} compilation target file
-   * @returns {String} full path to source map file
-   * @private
-   */
-  _getSourceMapFromFile (file) {
-    const fileBody = readFileSync(file).toString()
-    const sourceMapLineRE = /\/\/[#@] ?sourceMappingURL=([^\s'"]+)\s*$/mg
-    const results = fileBody.match(sourceMapLineRE)
-    if (results !== null) {
-      const sourceMap = results[results.length - 1].split('=')[1]
-      if (isAbsolute(sourceMap)) {
-        return sourceMap
-      } else {
-        const base = dirname(file)
-        return join(base, sourceMap)
-      }
-    }
-  }
-
-  /**
    * Returns the merged V8 process coverage.
    *
    * The result is computed from the individual process coverages generated
@@ -186,7 +164,7 @@ class Report {
           const ext = extname(f)
           if (ext === '.js' || ext === '.ts' || ext === '.mjs') {
             const stat = statSync(f)
-            const sourceMap = this._getSourceMapFromFile(f)
+            const sourceMap = getSourceMapFromFile(f)
             if (sourceMap !== undefined) {
               this.sourceMapCache[`file://${fullPath}`] = { data: JSON.parse(readFileSync(sourceMap).toString()) }
             }
